Outset handles so they overlap the border and not the content
Before, the resize handles overlapped the padding and a bit of the content,
now they overlap the padding and border, as intended.

I had to ditch `contain: paint` to get the handles to be interactive over the border part.
Looks like it doesn't affect just painting, and clips to the inner edge of the border, unlike the MDN docs say:
"And if a descendant overflows the containing element's bounds, then that descendant will be clipped to the containing element's border-box."
It seems to be clipped to the padding-box, at least if we're going by the definition used by `background-clip`