[demo] Clean up, simplify, comment code, and fix remaining scroll jank
I wanted to have the dynamic content define its own space in the page, but with scroll jank issues manifesting when loading the page and even when dragging/resizing a window the first time after page load in Firefox when the page is loaded scrolled down and zoomed in, I see now that it's not worth it. It was also a lot of code once I tried to work around these issues, and I basically ended up statically defining heights of things anyways (what I wanted to avoid) - just via code instead of directly. This also clarifies the code a lot. So does adding comments to sections of the demo code.