Fix subpixel issues with menu button borders, and margin-bottom
This was the most overwrought CSS I've ever written.
It was an interesting exercise in taking on too many requirements, and it WORKED! But not adding an extra element is a silly constraint.

And it had some problems:
1. With non-integer width text, browsers normally snap borders to device pixels, but with this solution the borders sometimes became blurry in Chrome, and in Firefox, stayed crisp but occasionally disappeared (depending on the exact widths of text rendered in a particular font)
2. It was missing the margin-bottom of 1px; the menu popups were still aligned correctly relatively, but the bottom shadow/highlight got cut off when using it in an application (specifically JS Paint).
3. So. Much. Code! Yeah it's way too complex for what it does.

Anyways while I am proud of my delicate border-balancing monstrosity (it does show my CSS ability), this commit on the other hand displays my restraint and maturity as a developer, as I have no qualms with binning this contrived code and simply adding a span.