Fix taskbarless minimization layout
The positioning of windows when minimized without a taskbar was broken
in https://github.com/1j01/os-gui/commit/82c4f2896990
when adding conversion to strings to resolve type checking errors.
I think I was exhausted, judging by the "This is exhausting." in the
commit message.

Actually, the type errors were not due to these being numbers, but
due to `titlebar_height` being possibly `undefined`.

Also, I've started adding tests for the windowing system.