[meta] This changes the error message...
I tried changing `cy.then(() => { expect(win.document.activeElement).to.equal(win.document.body); });` to `cy.focused().should('match', 'body');`, and similar, and it worked for 3/4 cases, but not this one. You might think it's because it's the body, and it considers the body to not be focusable, but another of the 4 cases is also looking for the body, and it works. It's perplexing.

If I comment out this line before it:
```js
				cy.get('#input').should('have.focus');
				cy.get('body').click({ force: true });
				// cy.get('#input').should('not.have.focus');
				cy.focused().should('match', 'body');
```
I get:
```
     AssertionError: Timed out retrying after 4000ms: expected '<input#input>' to match 'body'
```
whereas if I uncomment it... actually I'm getting the same error.
Before I got a different error... OK I got it again:
```
     AssertionError: Timed out retrying after 4000ms: Expected to find element: `focused`, but never found it.
```
It seems to depend on whether I mark the test with `.only`, which is "Only available when invoked via the mocha CLI", so it's not actually working for me, as I'm running via the cypress CLI.
Jeez, two pretty inexplicable error messages, inexplicably tied to a test framework feature that doesn't otherwise apply.