Start separating jQuery from $Window API
This is the first major step towards removing jQuery from os-gui's
windowing implementation. (I already removed it from the menu code.)

I used a `Proxy` object to overlay the jQuery methods onto the
`OSGUIWindow` interface methods, while defining the `OSGUIWindow`
interface in a separate object.
I could probably use `Object.assign()` for this, but I would have to
do it at the end of the code and make sure not to change properties
after construction (`icons` at least is stateful), and a `Proxy` will
allow me to introduce a deprecation notice for accessing jQuery methods.

I haven't changed the API yet, or introduced a non-jQuery based API.
I've just made it a bit less jQuery-based, or less entangled.

All of the tests pass.

Next up might be to convert the `OSGUIWindow` interface to a class,
and expose it as a new API, deprecating `$Window`.